# crates/markitai-cli/src/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Markitai service API",
    "version": "set from the build when served",
    "summary": "Convert files and URLs to Markdown, manage saved jobs and model settings.",
    "description": "The REST API of `markitai serve`. Loopback clients are trusted; other clients send the startup token as `Authorization: Bearer` (preferred) or `?token=`. Every error body is `{detail, code, reason}`; present text by `reason`, then `code`. See docs/serve.md, docs/service-settings.md and docs/provider-management.md for the contracts this document summarizes.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "/", "description": "The service that serves this document" }],
  "security": [{}, { "bearerAuth": [] }, { "tokenQuery": [] }],
...
  "paths": {
...
    "/api/settings/llm/detected": {
      "get": {
        "tags": ["providers"],
        "operationId": "listDetectedModels",
        "summary": "Quick-add candidates from provider keys in the service's environment",
        "responses": {
          "200": { "description": "Candidates", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/DetectedModel" } } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm/model-discovery": {
      "post": {
        "tags": ["providers"],
        "operationId": "discoverModels",
        "summary": "List a provider's models with the given or saved connection",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DiscoveryRequest" } } } },
        "responses": {
          "200": { "description": "Discovery result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DiscoveryResult" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm/test": {
      "post": {
        "tags": ["providers"],
        "operationId": "testModel",
        "summary": "One short request to exactly one deployment",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProbeRequest" } } } },
        "responses": {
          "200": { "description": "Test outcome; a provider failure is `ok: false`", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProbeResult" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "The startup token (printed at start, or MARKITAI_SERVE_TOKEN). Loopback peers need none." },
      "tokenQuery": { "type": "apiKey", "in": "query", "name": "token", "description": "The startup token as a query parameter; accepted for compatibility, but it can end up in logs and history, so prefer the header." },
      "ticketQuery": { "type": "apiKey", "in": "query", "name": "ticket", "description": "A single-use download ticket from POST /api/download-tickets, valid for one GET of the path it was issued for within 60 seconds." }
    },
...
    "responses": {
      "Error": { "description": "Refusal", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["detail", "code", "reason"],
        "properties": {
          "detail": { "description": "The service's English message, or an object for settings conflicts that carry `detail.code`", "oneOf": [{ "type": "string" }, { "type": "object" }] },
          "code": { "type": "string", "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "method_not_allowed", "conflict", "payload_too_large", "invalid_request", "rate_limited", "unavailable", "server_error"] },
          "reason": {
            "type": "string",
            "description": "Stable machine code of the specific cause. The list is the set the service writes; new codes may be added.",
            "examples": ["ambiguous_legacy_model_name", "archive_too_large", "batch_pending", "config_changed", "config_missing", "config_too_large", "discovery_timeout", "empty_job", "enhancement_failed", "file_not_found", "file_too_large", "form_field_too_large", "history_empty", "host_not_allowed", "internal_error", "invalid_allowed_host", "invalid_json", "invalid_multipart", "invalid_options", "invalid_provider_request", "invalid_retry_body", "invalid_settings_request", "invalid_ticket_path", "invalid_url", "invalid_urls", "item_busy", "item_not_found", "job_id_collision", "job_not_found", "job_not_running", "job_running", "json_required", "llm_unavailable", "method_not_allowed", "no_output", "no_source", "not_retryable", "nothing_to_stop", "origin_not_allowed", "output_conflict", "output_identity_conflict", "persistence_failed", "provider_busy", "provider_unavailable", "remote_url_forbidden", "request_too_large", "result_too_large", "result_unavailable", "route_not_found", "settings_durability_unknown", "settings_entry_missing", "settings_forbidden", "settings_io_failed", "settings_read_only", "shutting_down", "stale_revision", "ticket_invalid", "token_required", "too_many_items", "too_many_tickets", "unknown_preset", "unsupported_url_scheme", "upload_missing"]
          }
        }
      },
...
      "DetectedModel": {
        "type": "object",
        "required": ["provider", "model", "label", "requires_api_key"],
        "properties": {
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "label": { "type": "string" },
          "requires_api_key": { "type": "boolean" }
        }
      },
      "DiscoveryRequest": {
        "type": "object",
        "required": ["provider"],
        "additionalProperties": false,
        "properties": {
          "provider": { "type": "string" },
          "provider_id": { "type": "string", "description": "Use a saved connection's credentials" },
          "deployment_id": { "type": "string", "description": "Use a saved deployment's credentials" },
          "api_key": { "type": "string" },
          "api_base": { "type": "string" },
          "refresh": { "type": "boolean" }
        }
      },
      "DiscoveryResult": {
        "type": "object",
        "required": ["provider", "status", "models"],
        "properties": {
          "provider": { "type": "string" },
          "status": { "type": "string", "enum": ["ok", "partial", "unavailable"] },
          "source": { "type": "string" },
          "authoritative": { "type": "boolean" },
          "cached": { "type": "boolean" },
          "stale": { "type": "boolean" },
          "models": { "type": "array", "items": { "type": "object", "required": ["model", "label", "supports_vision"], "properties": { "model": { "type": "string" }, "label": { "type": "string" }, "supports_vision": { "type": "boolean" } } } },
          "detail": { "type": "string" }
        }
      },
      "ProbeRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Name a saved deployment, or a model with optional credentials.",
        "properties": {
          "deployment_id": { "type": "string" },
          "model_name": { "type": "string" },
          "model": { "type": "string" },
          "api_key": { "type": "string" },
          "api_base": { "type": "string" }
        }
      },
      "ProbeResult": {
        "type": "object",
        "required": ["ok", "detail"],
        "properties": { "ok": { "type": "boolean" }, "detail": { "type": "string" } }
      }
    }
  }
}